Add various SQL queries to analyze data
Also update models.sql to record prices as BIGINT to allow summation of two large numbers.

+++ src/db/models.sql
@@ -18,7 +18,7 @@ $$;
 CREATE TABLE IF NOT EXISTS fact_item (
   id INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
   item_id INT NOT NULL,
-  price INT NOT NULL,
+  price BIGINT NOT NULL,
   timestamp TIMESTAMPTZ NOT NULL,
   type item_type NOT NULL,
